refactor(steps): move feature generation/selection
