+++ ibis_ml/steps/__init__.py
@@ -6,9 +6,9 @@ from ibis_ml.steps._encode import (
     OneHotEncode,
     TargetEncode,
 )
-from ibis_ml.steps._feature_engineering import CreatePolynomialFeatures
-from ibis_ml.steps._feature_selection import DropZeroVariance
+from ibis_ml.steps._generate_features import CreatePolynomialFeatures
 from ibis_ml.steps._impute import FillNA, ImputeMean, ImputeMedian, ImputeMode
+from ibis_ml.steps._select_features import DropZeroVariance
 from ibis_ml.steps._standardize import ScaleMinMax, ScaleStandard
 from ibis_ml.steps._temporal import ExpandDate, ExpandDateTime, ExpandTime
